Update Compare_2_Files.py
Enhancements made to make The load list function for loading the Master file more flexible. The Master (or Lifer) file downloaded as a CSV file can be straightaway used. No need to just have one column with only the Species list.

Added flexibility is that the position of the Species column is determined at runtime by examining the Header row of the Lifer file.